chore: replace Pika with Microbundle
Pika doesn't seem to be actively maintained and is not compatible with
the latest version of TypeScript. Microbundle seems even easier to work
with than Pika and _is_ actively maintained.

BREAKING CHANGE: The package output is slightly different now, with a different configuration of the `package.json` pointers to different builds of the library. Because of this, the bundler consuming this library might pull in something different than before. In case that ends up being a "breaking" change, this commit is being marked as such, though in reality this should be safe for any consumers.